n=float(input()) a=3.14*n**2 print('{:.2f}'.format(a))

+++ Area_of_a_circle.py
@@ -1,3 +1,3 @@
-n=int(input())
+n=float(input())
 a=3.14*n**2
 print('{:.2f}'.format(a))
